refactor: Separate DashboardInvitations and Members' view and logic
